# serviflex-server/docs/swagger.yaml
definitions:
...
  models.Login:
    properties:
      email:
        type: string
      senha:
        type: string
    type: object
...
  models.Usuario:
    properties:
      email:
        type: string
      id:
        type: string
      nome:
        type: string
      senha:
        type: string
      tipo:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /cadastro:
    post:
      consumes:
      - application/json
      description: Cadastra novo cliente ou profissional no Firestore
      parameters:
      - description: Dados do usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/models.Usuario'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Usuario'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cadastro de usuário
      tags:
      - Autenticação
...
  /login:
    post:
      consumes:
      - application/json
      description: Autentica um usuário com email e senha
      parameters:
      - description: Email e Senha
        in: body
        name: credenciais
        required: true
        schema:
          $ref: '#/definitions/models.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Usuario'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login de usuário
      tags:
      - Autenticação
...
swagger: "2.0"